Fix constant type lost on merge

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/eip4844/ExecutionPayloadSchemaEip4844.java
@@ -104,7 +104,7 @@ public class ExecutionPayloadSchemaEip4844
         namedSchema(
             WITHDRAWALS,
             SszListSchema.create(
-                new WithdrawalSchema(), specConfig.getMaxWithdrawalsPerPayload().longValue())));
+                new WithdrawalSchema(), specConfig.getMaxWithdrawalsPerPayload())));
     this.defaultExecutionPayload = createFromBackingNode(getDefaultTree());
   }
